+++ src/components/SecurityCodeInput/SecurityCodeInput.tsx
@@ -4,23 +4,26 @@ import { useFormContext } from 'react-hook-form';
 
 import { digitInputTransformer } from '../../utils';
 import { FormControl } from '../Form';
-import { getKeyUpHandler } from './utils';
+import { getOnChangeHandler } from './utils';
 
 export interface SecurityCodeInputProps<NextElement> {
   className?: string;
+  inputRef: RefObject<HTMLInputElement>;
   inputClassName?: string;
   name: string;
-  nextFocusRef: RefObject<NextElement>;
+  nextFocusRef?: RefObject<NextElement>;
 }
 
 export const SecurityCodeInput = <NextElement extends HTMLElement>({
   className,
+  inputRef,
   inputClassName,
   name,
   nextFocusRef,
 }: SecurityCodeInputProps<NextElement>) => {
   const { setValue } = useFormContext();
   const digit1Ref = useRef<HTMLInputElement>(null);
+  const firstInputRef = inputRef ?? digit1Ref;
   const digit2Ref = useRef<HTMLInputElement>(null);
   const digit3Ref = useRef<HTMLInputElement>(null);
   const digit4Ref = useRef<HTMLInputElement>(null);
@@ -30,8 +33,8 @@ export const SecurityCodeInput = <NextElement extends HTMLElement>({
     <div className={classNames('flex gap-2', className)}>
       <FormControl
         hideErrorMessage
-        inputRef={digit1Ref}
-        onKeyUp={getKeyUpHandler(digit2Ref)}
+        inputRef={firstInputRef}
+        onChange={getOnChangeHandler(digit2Ref)}
         onPaste={event => {
           const data = event.clipboardData.getData('text');
           const digits = data.split('');
@@ -41,7 +44,7 @@ export const SecurityCodeInput = <NextElement extends HTMLElement>({
           setValue(`${name}.digit4`, digits[3] ?? '');
           setValue(`${name}.digit5`, digits[4] ?? '');
           setValue(`${name}.digit6`, digits[5] ?? '');
-          nextFocusRef.current?.focus();
+          nextFocusRef?.current?.focus();
         }}
         inputClassName={inputClassName}
         name={`${name}.digit1`}
@@ -50,7 +53,7 @@ export const SecurityCodeInput = <NextElement extends HTMLElement>({
       <FormControl
         hideErrorMessage
         inputRef={digit2Ref}
-        onKeyUp={getKeyUpHandler(digit3Ref)}
+        onChange={getOnChangeHandler(digit3Ref)}
         inputClassName={inputClassName}
         name={`${name}.digit2`}
         transform={digitInputTransformer}
@@ -58,7 +61,7 @@ export const SecurityCodeInput = <NextElement extends HTMLElement>({
       <FormControl
         hideErrorMessage
         inputRef={digit3Ref}
-        onKeyUp={getKeyUpHandler(digit4Ref)}
+        onChange={getOnChangeHandler(digit4Ref)}
         inputClassName={inputClassName}
         name={`${name}.digit3`}
         transform={digitInputTransformer}
@@ -66,7 +69,7 @@ export const SecurityCodeInput = <NextElement extends HTMLElement>({
       <FormControl
         hideErrorMessage
         inputRef={digit4Ref}
-        onKeyUp={getKeyUpHandler(digit5Ref)}
+        onChange={getOnChangeHandler(digit5Ref)}
         inputClassName={inputClassName}
         name={`${name}.digit4`}
         transform={digitInputTransformer}
@@ -74,7 +77,7 @@ export const SecurityCodeInput = <NextElement extends HTMLElement>({
       <FormControl
         hideErrorMessage
         inputRef={digit5Ref}
-        onKeyUp={getKeyUpHandler(digit6Ref)}
+        onChange={getOnChangeHandler(digit6Ref)}
         inputClassName={inputClassName}
         name={`${name}.digit5`}
         transform={digitInputTransformer}
@@ -82,7 +85,7 @@ export const SecurityCodeInput = <NextElement extends HTMLElement>({
       <FormControl
         hideErrorMessage
         inputRef={digit6Ref}
-        onKeyUp={getKeyUpHandler(nextFocusRef)}
+        onChange={getOnChangeHandler(nextFocusRef)}
         inputClassName={inputClassName}
         name={`${name}.digit6`}
         transform={digitInputTransformer}