+++ src/components/SecurityCodeInput/constants.ts
@@ -1,0 +1,8 @@
+export const defaultSecurityCodeInputValues = {
+  digit1: '',
+  digit2: '',
+  digit3: '',
+  digit4: '',
+  digit5: '',
+  digit6: '',
+};