+++ src/components/SecurityCodeInput/index.ts
@@ -1,1 +1,3 @@
 export * from './SecurityCodeInput';
+export * from './constants';
+export * from './schema';